Provide more info in error messages
The response['detail'] that is passed into the error classes is
sometimes blank, so the standard #to_s or e.message doesn't provide
any more detail about the error message. This changes it to use
response['title'], which every error message passed back from
Responsys contains, and adds response['detail'] if it's available.